feat: Introduces Vitest and Locations MFE
Sets up Vitest and React Testing Library for robust unit and integration testing. This includes new test scripts, dependencies, and initial tests for Header and Footer components to ensure code quality and maintainability.

Adds a new 'Locations' page that dynamically loads a Micro-Frontend (MFE). This enhances the application's features by providing a dedicated section for locations, improving modularity and scalability. The welcome page is updated to offer navigation to this new MFE.